+++ R/tlf-env.R
@@ -7,10 +7,10 @@ tlfEnv$packageName <- "tlf"
 
 #' @title LegendPositions
 #' @import ospsuite.utils
-#' @export
 #' @description
 #' List of all available legend positions
 #' @family enum helpers
+#' @export
 LegendPositions <- enum(c(
   "none",
   "insideTop",
@@ -32,6 +32,59 @@ LegendPositions <- enum(c(
 ))
 
 tlfEnv$defaultLegendPosition <- LegendPositions$outsideRight
+
+#' @title TagPositions
+#' @import ospsuite.utils
+#' @description
+#' List of all available tag positions in a plot grid.
+#' @family enum helpers
+#' @export
+TagPositions <- enum(
+  c(
+    "topLeft" = "topleft",
+    "top" = "top",
+    "topRight" = "topright",
+    "left" = "left",
+    "right" = "right",
+    "bottomLeft" = "bottomleft",
+    "bottom" = "bottom",
+    "bottomRight" = "bottomright"
+  )
+)
+
+tlfEnv$defaultTagPosition <- TagPositions$topLeft
+
+#' @title HorizontalJustification
+#' @import ospsuite.utils
+#' @description
+#' List of all available horizontal justifications for plot annotation text.
+#' @family enum helpers
+#' @export
+HorizontalJustification <- enum(
+  c(
+    "left" = 0,
+    "middle" = 0.5,
+    "right" = 1
+  )
+)
+
+tlfEnv$defaultHorizontalJustification <- HorizontalJustification$left
+
+#' @title VerticalJustification
+#' @import ospsuite.utils
+#' @description
+#' List of all available vertical justifications for plot annotation text.
+#' @family enum helpers
+#' @export
+VerticalJustification <- enum(
+  c(
+    "bottom" = 0,
+    "middle" = 0.5,
+    "top" = 1
+  )
+)
+
+tlfEnv$defaultVerticalJustification <- VerticalJustification$bottom
 
 #' @title setDefaultLegendPosition
 #' @description Set default legend position of tlf environment